refactor: rename service for trip (DD-9)
